requirements.txt
pytest==9.0.3
pytest-cov
pytest-playwright==0.7.2
ruff== 0.15.12